refactor(snapshot-versions): move snapshots_active into snapshot_versions package (architect MEDIUM, PR #60)
