+++ backend/services/snapshot_versions/__init__.py
@@ -1,1 +1,8 @@
 """Snapshot Version lifecycle services."""
+
+from services.snapshot_versions.active import (
+    ActiveReleaseMissingError,
+    get_active_release_id,
+)
+
+__all__ = ["ActiveReleaseMissingError", "get_active_release_id"]
